Implement user use case with improved error handling and logging
