+++ applications/app-service/src/main/resources/application.yaml
@@ -14,6 +14,9 @@ spring:
     password: "pass12345"
     locations: "classpath:db/migration"
     enabled: true
+  output:
+    ansi:
+      enabled: ALWAYS
 
 adapters:
   r2dbc:
@@ -46,4 +49,3 @@ springdoc:
       description: "This is the API for Crediya Auth Microservice"
   swagger-ui:
     path: /swagger-ui.html
-
