+++ domain/model/src/main/java/co/com/pragma/model/logs/gateways/LoggerPort.java
@@ -2,9 +2,11 @@ package co.com.pragma.model.logs.gateways;
 
 public interface LoggerPort {
 
+    void debug(String message, Object... args);
+
     void info(String message, Object... args);
 
     void warn(String message, Object... args);
 
-    void error(String message, Throwable throwable);
+    void error(String message, Object... args);
 }
