--- domain/usecase/src/main/java/co/com/pragma/usecase/role/RoleUseCase.java
@@ -1,6 +1,0 @@
-package co.com.pragma.usecase.role;
-
-import lombok.RequiredArgsConstructor;
-@RequiredArgsConstructor
-public class RoleUseCase {
-}
